Removed unnessary patch in spk/salt-minion/src/service-setup.sh. Added failsafe logic for symlink creation

+++ spk/salt-master/src/service-setup.sh
@@ -17,13 +17,8 @@ service_postinst ()
     wheelhouse=${SYNOPKG_PKGDEST}/share/wheelhouse
     ${SYNOPKG_PKGDEST}/env/bin/pip install --no-deps --no-index --upgrade --force-reinstall -f ${wheelhouse} ${wheelhouse}/*.whl
 
-    # Patch rsax931.py file to find libcrypto lib
-    # (Rely on patch util bundled with python3's busybox)
-    ${SYNOPKG_PKGDEST}/bin/patch ${SYNOPKG_PKGDEST}/env/lib/python3.8/site-packages/salt/utils/rsax931.py -i ${SYNOPKG_PKGDEST}/share/rsax931.py.patch
-
     # Prepare salt-master config in /var/packages/salt-minion/
-    ln -s /var/packages/${SYNOPKG_PKGNAME}/etc ${SYNOPKG_PKGDEST}/env/etc
-    test -d ${SYNOPKG_PKGDEST}/env/etc/salt || install -m 755 -o sc-${SYNOPKG_PKGNAME} -g ${SYNOPKG_PKGNAME} -d ${SYNOPKG_PKGDEST}/env/etc/salt
+    test -L ${SYNOPKG_PKGDEST}/env/etc || ln -s /var/packages/${SYNOPKG_PKGNAME}/etc ${SYNOPKG_PKGDEST}/env/etc
     test -d ${SYNOPKG_PKGDEST}/env/etc/salt/master.d || install -m 755 -o sc-${SYNOPKG_PKGNAME} -g ${SYNOPKG_PKGNAME} -d ${SYNOPKG_PKGDEST}/env/etc/salt/master.d
     test -f ${SYNOPKG_PKGDEST}/env/etc/salt/master || install -m 644 -o sc-${SYNOPKG_PKGNAME} -g ${SYNOPKG_PKGNAME} ${SYNOPKG_PKGDEST}/share/master ${SYNOPKG_PKGDEST}/env/etc/salt/master
     test -f ${SYNOPKG_PKGDEST}/env/etc/salt/master.d/02_pidfile.conf || echo "pidfile: ${PID_FILE}" > ${SYNOPKG_PKGDEST}/env/etc/salt/master.d/02_pidfile.conf
